+++ app/Http/Controllers/ProductController.php
@@ -15,6 +15,8 @@ class ProductController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Product::class);
+
         $products =Product::paginate(5);
         // dd($products);
         // $categories = Category::all();
@@ -33,6 +35,8 @@ class ProductController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Product::class);
+
         $categories=Category::get();
         $param = [
             'categories' => $categories
@@ -100,6 +104,8 @@ class ProductController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('view', Product::class);
+
         $productshow = Product::findOrFail($id);
         $param =[
             'productshow'=>$productshow,
@@ -117,6 +123,8 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('update', Product::class);
+
         $product = Product::find($id);
         $categories=Category::get();
         $param = [
@@ -207,17 +215,20 @@ class ProductController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('forceDelete', Product::class);
         $products=Product::onlyTrashed()->findOrFail($id);
         $products->forceDelete();
 
     }
     public  function trash(){
+        $this->authorize('viewtrash', Product::class);
         $products = Product::onlyTrashed()->get();
         $param = ['products'    => $products];
         return view('admin.product.trash', $param);
     }
 
     public  function softdeletes($id){
+        $this->authorize('delete', Product::class);
         date_default_timezone_set("Asia/Ho_Chi_Minh");
         $product = Product::findOrFail($id);
         $product->deleted_at = date("Y-m-d h:i:s");
@@ -230,6 +241,7 @@ class ProductController extends Controller
     }
 
     public function restoredelete($id){
+        $this->authorize('restore', Product::class);
         $product=Product::withTrashed()->where('id', $id);
         $product->restore();
         $notification = [
